Reorganizing: moved class objects
I moved all the different classes into the Driver.java class so they can all interact with eachother. An oversight in my original organization of the class structures put separate instances of communicating objects into their private fields, but this would make the VM impossible. they need to be in the same class so all the objects can communicate with eachother.